+++ pages/modules_builder.php
@@ -1,0 +1,8 @@
+<?php
+
+/**
+ * Builder - Buildermodul (Full Builder)
+ */
+
+$builderModulesView = 'full';
+include rex_path::addon('builder', 'pages/modules.php');