+++ pages/modules_single.php
@@ -1,0 +1,8 @@
+<?php
+
+/**
+ * Builder - Einzelmodule
+ */
+
+$builderModulesView = 'single';
+include rex_path::addon('builder', 'pages/modules.php');